rename userAuth -> auth
