+++ .vscode/settings.json
@@ -13,7 +13,6 @@
         "sirupsen",
         "Usecase",
         "usecases",
-        "userauth",
         "Warnf"
     ]
 }
