--- modules/userauth/userauth.go
@@ -1,10 +1,0 @@
-package userauth
-
-import (
-	"template-golang/modules/userauth/middlewares"
-)
-
-// Dependencies contains all dependencies for the module
-type UserAuth struct {
-	Handler middlewares.UserAuthMiddleware
-}
