+++ server/wire.go
@@ -7,8 +7,8 @@ package server
 import (
 	"template-golang/config"
 	"template-golang/database"
+	"template-golang/modules/auth"
 	"template-golang/modules/cockroach"
-	"template-golang/modules/userauth"
 
 	"github.com/google/wire"
 )
@@ -23,7 +23,7 @@ var ProviderSet = wire.NewSet(
 
 	// modules
 	cockroach.ProviderSet,
-	userauth.ProviderSet,
+	auth.ProviderSet,
 )
 
 func Wire() (Server, error) {
